Move Metric to types.rs
